treewide: plug memleak: case 2.2.2: for_each_child_of_node() (v6.12 prep)
Case 2.2.2 from below bug:

    If returning or breaking early from within
    for_each_child_of_node(), a reference to the current device_node
    remains and needs to be dropped, either right away, or later if
    used past that loop to avoid a memleak.

In this case, the np is not needed past the loop and code can be
simplified by switching to for_each_child_of_node_scoped() which avoids
the need for manual cleanup in early exits from the loop.

This commit plugs memleaks in both normal and error code path.

Test: TH
Bug: 370679843
Change-Id: Id8cb40d1d4fe271cc92be30aa654a7c0a8bf36d7
Signed-off-by: André Draszik <draszik@google.com>
Signed-off-by: Will McVicker <willmcvicker@google.com>